Bug fix: allows number input after performing an operation
Previously, after user performs an operation, if the user immediately
clicks number buttons, the numbers will be assigned to num2 while
operator is null. This will cause the calculator to be out of sequence
and eventually result in it trying to make a calculation on num1 and
num2 without an operator, resulting in an error. Now, if the user clicks
number buttons after an operation, the number input will override num1.
If the user clicks an operator button after an operation, num1 will
be the result of the previous operation and the operator button will be
stored in formula object.