Adds scientific notation for numbers larger than 9 significant digits
This is mainly used for the result of a calculation. If the result has
more than 9 significant digits and is not a continuous decimal number,
then calcScreen will show scientific notation (program will still store
the full value of the number). BACKSPACE functionality is restored to
remove 1 digit at a time even when a number is displayed with
scientific notation.